backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
supabase==2.9.1
python-dotenv==1.0.1
httpx==0.27.0
openai==2.32.0
apscheduler==3.10.4
python-multipart==0.0.12
PyJWT[crypto]==2.9.0
cryptography==43.0.1
